T-9810773-Create-an-adaptive-layout-of-playing-field: refactoring
