+++ src/App.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './App.css';
+import { MainPage } from './components/MainPage';
+
+function App() {
+  return (
+    <div>
+      <MainPage />
+    </div>
+  );
+}
+
+export default App;
