--- src/components/App.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { MainPage } from './index';
-import './App.css';
-
-function App() {
-  return (
-    <div>
-      <MainPage />
-    </div>
-  );
-}
-
-export default App;
