+++ src/components/MainPage/CommonGameArea/CardArea/CardArea.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { GameCardAreaWrapper } from './CardAreaStyled';
+
+export function CardArea() {
+  return (
+    <GameCardAreaWrapper>
+      something
+    </GameCardAreaWrapper>
+  );
+}
