+++ src/components/MainPage/CommonGameArea/CardArea/CardAreaStyled.tsx
