+++ src/components/MainPage/CommonGameArea/CardArea/index.ts
@@ -1,0 +1,1 @@
+export * from './CardArea';
