+++ src/components/MainPage/CommonGameArea/CardsPanel/CardsPanel.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { CardsPanelArea } from './CardsPanelStyled';
+
+export function CardsPanel() {
+  return (
+    <CardsPanelArea>
+      Cards Panel
+    </CardsPanelArea>
+  );
+}
