+++ src/components/MainPage/CommonGameArea/CardsPanel/CardsPanelStyled.tsx
