+++ src/components/MainPage/CommonGameArea/CardsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './CardsPanel';
