+++ src/components/MainPage/CommonGameArea/CommonGameArea.styles.tsx
