+++ src/components/MainPage/CommonGameArea/GameBoard/GameAreaWrapper/GameAreaWrapper.tsx
