+++ src/components/MainPage/CommonGameArea/GameBoard/GameAreaWrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './GameAreaWrapper';
