+++ src/components/MainPage/CommonGameArea/GameBoard/GameBoardPart/GameBoardSide.types.ts
@@ -1,0 +1,8 @@
+export interface Element {
+  id: string;
+  value: number;
+}
+
+export interface GameBoardProps {
+  elements: Element[];
+}
