+++ src/components/MainPage/CommonGameArea/GameBoard/GameBoardPart/index.ts
@@ -1,0 +1,1 @@
+export * from './GameBoardSide';
