+++ src/components/MainPage/CommonGameArea/GameBoard/index.ts
@@ -1,0 +1,1 @@
+export * from './GameBoard';
