+++ src/components/MainPage/CommonGameArea/UserCards/UserCardsWrapperStyled.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { colors } from '../../../../theme/variables';
+
+export const UserCardsWrapperStyled = styled.div`
+    grid-row: 2 / 3;
+    grid-column: 1 / 2;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 20vh;
+    background-color: ${colors.cardsListBackgroundColor};
+    overflow: auto;
+`;
