+++ src/components/MainPage/CommonGameArea/UserCards/index.ts
@@ -1,0 +1,1 @@
+export * from './UserCards';
