+++ src/components/MainPage/CommonGameArea/index.ts
@@ -1,0 +1,1 @@
+export * from './CommonGameArea';
