+++ src/components/MainPage/ScorePanel/ScorePanel.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { ScorePanelStyled } from './ScorePanelStyled';
+
+export function ScorePanel() {
+  return (
+    <ScorePanelStyled>score panel</ScorePanelStyled>
+  );
+}
