+++ src/components/MainPage/ScorePanel/index.ts
@@ -1,0 +1,1 @@
+export * from './ScorePanel';
