+++ src/components/Shared/Card/Card.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Card } from './CardStyled';
+
+export function CommonGameArea() {
+  return (
+    <Card />
+  );
+}
