--- src/components/mainPage/commonGameArea/cardsList/index.ts
@@ -1,1 +1,0 @@
-export * from './cardsList';
