--- src/components/mainPage/commonGameArea/cardsPanel/index.ts
@@ -1,1 +1,0 @@
-export * from './cardsPanel';
