--- src/components/mainPage/commonGameArea/gameArea/gameArea.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { GameAreaSide } from './gameAreaSides';
-import { GameAreaWrapper } from './gameAreaWrapper';
-
-const elementsEnemy = [
-  { id: 'enemyFirst', value: 1 },
-  { id: 'enemySecond', value: 2 },
-  { id: 'enemyThird', value: 3 },
-];
-const elementsFriendly = [
-  { id: 'friendlyFirst', value: 4 },
-  { id: 'friendlySecond', value: 5 },
-  { id: 'friendlyThird', value: 6 },
-];
-
-export function GameArea() {
-  return (
-    <GameAreaWrapper>
-      <GameAreaSide elements={elementsEnemy} />
-      <GameAreaSide elements={elementsFriendly} />
-    </GameAreaWrapper>
-  );
-}
