--- src/components/mainPage/commonGameArea/gameArea/gameAreaSides/gameAreaSidesStyles/index.ts
@@ -1,1 +1,0 @@
-export * from './gameAreaSidesStyles';
