--- src/components/mainPage/commonGameArea/gameArea/gameAreaSides/index.ts
@@ -1,1 +1,0 @@
-export * from './gameAreaSides';
