--- src/components/mainPage/commonGameArea/gameArea/gameAreaWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './gameAreaWrapper';
