--- src/components/mainPage/commonGameArea/gameArea/index.ts
@@ -1,1 +1,0 @@
-export * from './gameArea';
