--- src/components/mainPage/commonGameArea/gameCardArea/gameCardArea.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { GameCardAreaWrapper } from './gameCardAreaStyles';
-
-export function GameCardArea() {
-  return <GameCardAreaWrapper />;
-}
