--- src/components/mainPage/commonGameArea/gameCardArea/index.ts
@@ -1,1 +1,0 @@
-export * from './gameCardArea';
