--- src/components/mainPage/commonGameArea/index.ts
@@ -1,1 +1,0 @@
-export * from './commonGameArea';
