--- src/components/mainPage/index.ts
@@ -1,1 +1,0 @@
-export * from './mainPage';
