--- src/components/mainPage/opponentsPanel/index.ts
@@ -1,1 +1,0 @@
-export * from './opponentsPanel';
