--- src/components/mainPage/opponentsPanel/opponentsPanel.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { OpponentPanel } from './opponentsPanelStyles';
-
-export function OpponentsPanel() {
-  return (
-    <OpponentPanel />
-  );
-}
